Ignore whitespace changes by default
This changes the way we retrieve the diff we use for evaluation such
that we can ignore whitespace changes by default. Rather than retrieving
the diff from the GitHub API like we used to previously, we now compute
it manually using `git diff`. Doing so allows us to pass custom
arguments to `git diff` that we use to ignore whitespace changes by
default.