Add ability to run this workflow in shadow mode
In "shadow mode", this workflow will compute a score even for users who
have not opted into the workflow, but it will not do any labeling or
commenting. This sets up a future feature in which this workflow
serializes the scores it computes in order to enable discriminatory
analysis between those who have opted into the workflow and those wh
have not.